Correct example of delete_namespaced_deployment
grace_period_seconds, orpan_descendants, and propagation_policy are part of the V1DeleteOptions body parameter.  Body is a required parameter, and if it is empty and prop. policy is set then prop policy is ignored, resulting in orphans.  If prop policy is set in the body then orphans are removed as expected.  Replaced example code with working code.