Rest: Warn about swallowed _request_timeout
If `_request_timeout` is neither an int, nor a 2-tuple, it is swallowed
without further notice which is a rather unfortunate because the level
developers would have to look for this issue is pretty deep.

This actually leads to confusion already, see
https://github.com/apache/airflow/pull/6643#issuecomment-575438090

While it would break backwards compatibility to raise an exception, we
should at least warn the developer.