Force Lease Expiration When Leader Exits
Currently, when the leader exits (say, after receiving a SIGINT),
the workers need to wait for its lease to expire before a leader
is re-elected. This patch mimics the behaviour of the Go Client implementation
of using ctx.Done() by capturing the SIGINT and forcing the expiration date to a
past date and also sets the acquire_time to None to start the leader election.